refactor: a little cleanup
