+++ packages/example/package.json
@@ -11,7 +11,7 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@celo-tools/use-contractkit": "^1.0.0-beta.14-dev",
+    "@celo-tools/use-contractkit": "1.0.0-beta.14-dev",
     "@ethersproject/providers": "^5.3.0",
     "autoprefixer": "^10.2.4",
     "next": "11.0.1",
